Fix iBorrowPositionToken for timestamps as uint40

+++ packages/hardhat/contracts/interfaces/IBorrowPositionToken.sol
@@ -8,15 +8,15 @@ interface IBorrowPositionToken is IERC721 {
     function mint(address account, DataTypes.BorrowData calldata borrowData) external returns (uint);
     function debt(uint256 tokenId) external returns (uint);
 //    function interest(uint256 tokenId) external returns (uint256);
-    function interestSinceTimestamp(uint256 tokenId, uint timestamp) external returns (uint256);
+    function interestSinceTimestamp(uint tokenId, uint40 timestamp) external returns (uint256);
 //    function updateSnapshot(uint256 tokenId, uint newDebt) external;
     function getEndDate(uint tokenId) external returns (uint40);
     function getCollateral(uint tokenId) external returns (uint256);
-    function getStartDate(uint tokenId) external view returns (uint256);
+    function getStartDate(uint tokenId) external view returns (uint40);
     function burn(uint256 tokenId) external;
     function getTokensByTimestamp(uint40 _timestamp) external returns (uint[] calldata);
     function getLoan(uint tokenId) external returns (DataTypes.BorrowData memory);
     function getInterest(uint tokenId) external view returns (uint256);
-    function partialRepayLoan(uint256 tokenId, uint256 repaymentAmount, uint lastSnapshotDate, address sender) external returns(uint principalReduction);
+    function partialRepayLoan(uint256 tokenId, uint256 repaymentAmount, uint40 lastSnapshotDate, address sender) external returns(uint principalReduction);
     function cleanUpByTimestamp(uint40 timestamp) external;
 }
